LOKI: Split power supply banks into separate setup files
This makes it easier to organize the UI by showing/hiding banks
and their channels.

Channel info is imported from the PS config file. NOTE: the device
creation (channels, in this case), however, must be done in the same
file. NICOS does not allow it to be imported from the PS config file.